requirements/local.txt
# Local development dependencies go here
-r base.txt

# django-debug-toolbar that works with Django 1.5+
django-debug-toolbar==1.4